Small fixes in response to PR comments
