+++ coilsnake/assets/modulelist.txt
@@ -21,7 +21,7 @@ eb.WindowGraphicsModule
 eb.BattleBgModule
 eb.CompressedGraphicsModule
 eb.EnemyModule
-eb.TilesetModule
 eb.TitleScreenModule
+eb.TilesetModule
 smb.TextModule
 common.LunarIpsCompatibilityModule
